+++ tests/PSRule.Tests/PSRule.Assert.Tests.ps1
@@ -43,6 +43,7 @@ Describe 'PSRule assertions' -Tag 'Assert' {
                 Version = '2.0.0'
                 CompareNumeric = 3
                 CompareArray = 1, 2, 3
+                CompareString = 'abc'
             }
             [PSCustomObject]@{
                 Name = 'TestObject2'
@@ -57,8 +58,23 @@ Describe 'PSRule assertions' -Tag 'Assert' {
                 Version = '1.0.0'
                 CompareNumeric = 0
                 CompareArray = @()
+                CompareString = ''
             }
         )
+
+        It 'In pre-conditions' {
+            $result = @($testObject | Invoke-PSRule -Path $ruleFilePath -Name 'Assert.Precondition' -Outcome All -WarningAction SilentlyContinue);
+            $result | Should -Not -BeNullOrEmpty;
+            $result.Length | Should -Be 2;
+
+            # Positive case
+            $result[0].Outcome | Should -Be 'Pass';
+            $result[0].TargetName | Should -Be 'TestObject1';
+
+            # Negative case
+            $result[1].Outcome | Should -Be 'None';
+            $result[1].TargetName | Should -Be 'TestObject2';
+        }
 
         It 'Complete' {
             $result = @($testObject | Invoke-PSRule -Path $ruleFilePath -Name 'Assert.Complete');
@@ -120,7 +136,7 @@ Describe 'PSRule assertions' -Tag 'Assert' {
             # Negative case
             $result[1].IsSuccess() | Should -Be $False;
             $result[1].TargetName | Should -Be 'TestObject2';
-            $result[1].Reason.Length | Should -Be 2;
+            $result[1].Reason.Length | Should -Be 3;
             $result[1].Reason | Should -BeLike "The value '*' was not > '*'.";
         }
 
@@ -136,7 +152,7 @@ Describe 'PSRule assertions' -Tag 'Assert' {
             # Negative case
             $result[1].IsSuccess() | Should -Be $False;
             $result[1].TargetName | Should -Be 'TestObject2';
-            $result[1].Reason.Length | Should -Be 2;
+            $result[1].Reason.Length | Should -Be 3;
             $result[1].Reason | Should -BeLike "The value '*' was not >= '*'.";
         }
 
@@ -218,7 +234,7 @@ Describe 'PSRule assertions' -Tag 'Assert' {
             # Negative case
             $result[0].IsSuccess() | Should -Be $False;
             $result[0].TargetName | Should -Be 'TestObject1';
-            $result[0].Reason.Length | Should -Be 2;
+            $result[0].Reason.Length | Should -Be 3;
             $result[0].Reason | Should -BeLike "The value '*' was not < '*'.";
         }
 
@@ -234,7 +250,7 @@ Describe 'PSRule assertions' -Tag 'Assert' {
             # Negative case
             $result[0].IsSuccess() | Should -Be $False;
             $result[0].TargetName | Should -Be 'TestObject1';
-            $result[0].Reason.Length | Should -Be 2;
+            $result[0].Reason.Length | Should -Be 3;
             $result[0].Reason | Should -BeLike "The value '*' was not <= '*'.";
         }